feat: refactor application entry point to use Application struct
The application entry point has been refactored to use the Application struct. This struct serves as the entry point for a Gazelle application. The `build` method has been added to the Application struct to handle the initialization of the struct and return a Result indicating success or failure. The `run` method has been moved to the Application struct and is now called on an instance of the struct.

The contents of the application module have been moved back into lib.rs (for now)